requirements.txt
sagemaker
python-dotenv